[INPLACE-175] 모바일 체크 hook 분리
