+++ frontend/src/components/Main/MapSection/index.tsx
@@ -1,22 +1,13 @@
-import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { FaChevronRight } from 'react-icons/fa6';
 import VisitPlace from '@/assets/images/visit-place.png';
 import { Paragraph } from '@/components/common/typography/Paragraph';
+import useIsMobile from '@/hooks/useIsMobile';
 
 export default function MapSection() {
   const navigate = useNavigate();
-  const [isMobile, setIsMobile] = useState(window.innerWidth <= 768);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth <= 768);
-    };
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
+  const isMobile = useIsMobile();
 
   const handleConfirmClick = () => {
     navigate('/map');
