+++ frontend/src/hooks/useIsMobile.ts
@@ -1,0 +1,21 @@
+import { useEffect, useState } from 'react';
+
+const MOBILE_WIDTH_THRESHOLD = 768;
+
+export default function useIsMobile() {
+  const [isMobile, setIsMobile] = useState(
+    typeof window !== 'undefined' ? window.innerWidth <= MOBILE_WIDTH_THRESHOLD : false,
+  );
+
+  useEffect(() => {
+    const checkMobile = () => {
+      setIsMobile(window.innerWidth <= MOBILE_WIDTH_THRESHOLD);
+    };
+
+    checkMobile();
+    window.addEventListener('resize', checkMobile);
+    return () => window.removeEventListener('resize', checkMobile);
+  }, []);
+
+  return isMobile;
+}
